+++ src/main/java/com/vtortsev/quizapp/controller/QuestionController.java
@@ -1,13 +1,10 @@
 package com.vtortsev.quizapp.controller;
 
-import com.vtortsev.quizapp.dto.CategoryDto;
-import com.vtortsev.quizapp.dto.CreateCategoryDto;
-import com.vtortsev.quizapp.dto.CreateQuestionDto;
+import com.vtortsev.quizapp.dto.createEntityDto.CreateQuestionDto;
 import com.vtortsev.quizapp.dto.QuestionDto;
 import com.vtortsev.quizapp.dto.mapper.QuestionMapper;
 import com.vtortsev.quizapp.entities.Question;
 import com.vtortsev.quizapp.service.QuestionService;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
